+++ src/serena/tools/file_tools.py
@@ -12,7 +12,6 @@ from fnmatch import fnmatch
 from pathlib import Path
 from typing import Literal
 
-from serena.code_editor import EditedFilePath
 from serena.tools import SUCCESS_RESULT, EditedFileContext, EditingToolWithDiagnostics, Tool, ToolMarkerOptional
 from serena.util.file_system import scan_directory
 from serena.util.text_utils import ContentReplacer, search_files
@@ -62,30 +61,27 @@ class CreateTextFileTool(EditingToolWithDiagnostics):
         :param content: the (appropriately encoded) content to write to the file
         :return: a message indicating success or failure
         """
-        # capturing diagnostics before the edit
-        edited_file_paths = [EditedFilePath(relative_path, relative_path)]
-        diagnostics_snapshot = self._capture_published_lsp_diagnostics_snapshot(edited_file_paths)
+        with self.DiagnosticsContext(self, relative_path) as diagnostics_context:
+            # validating the destination path
+            project_root = self.get_project_root()
+            abs_path = (Path(project_root) / relative_path).resolve()
+            will_overwrite_existing = abs_path.exists()
 
-        # validating the destination path
-        project_root = self.get_project_root()
-        abs_path = (Path(project_root) / relative_path).resolve()
-        will_overwrite_existing = abs_path.exists()
+            if will_overwrite_existing:
+                self.project.validate_relative_path(relative_path, require_not_ignored=True)
+            else:
+                assert abs_path.is_relative_to(self.get_project_root()), (
+                    f"Cannot create file outside of the project directory, got {relative_path=}"
+                )
 
-        if will_overwrite_existing:
-            self.project.validate_relative_path(relative_path, require_not_ignored=True)
-        else:
-            assert abs_path.is_relative_to(self.get_project_root()), (
-                f"Cannot create file outside of the project directory, got {relative_path=}"
-            )
+            # writing the file
+            abs_path.parent.mkdir(parents=True, exist_ok=True)
+            abs_path.write_text(content, encoding=self.project.project_config.encoding, newline=self.project.line_ending.newline_str)
+            answer = f"File created: {relative_path}."
+            if will_overwrite_existing:
+                answer += " Overwrote existing file."
 
-        # writing the file
-        abs_path.parent.mkdir(parents=True, exist_ok=True)
-        abs_path.write_text(content, encoding=self.project.project_config.encoding, newline=self.project.line_ending.newline_str)
-        answer = f"File created: {relative_path}."
-        if will_overwrite_existing:
-            answer += " Overwrote existing file."
-
-        return self._format_lsp_edit_result_with_new_diagnostics(answer, edited_file_paths, diagnostics_snapshot)
+            return diagnostics_context.format_result(answer)
 
 
 class ListDirTool(Tool):
@@ -220,19 +216,14 @@ class ReplaceContentTool(EditingToolWithDiagnostics):
         Performs the replacement, with additional options not exposed in the tool.
         This function can be used internally by other tools.
         """
-        # capturing diagnostics before the edit
-        edited_file_paths = [EditedFilePath(relative_path, relative_path)]
-        diagnostics_snapshot = self._capture_published_lsp_diagnostics_snapshot(edited_file_paths)
-
-        # applying the content replacement
-        self.project.validate_relative_path(relative_path, require_not_ignored=require_not_ignored)
-        with EditedFileContext(relative_path, self.create_code_editor()) as context:
-            original_content = context.get_original_content()
-            replacer = ContentReplacer(mode=mode, allow_multiple_occurrences=allow_multiple_occurrences)
-            updated_content = replacer.replace(original_content, needle, repl)
-            context.set_updated_content(updated_content)
-
-        return self._format_lsp_edit_result_with_new_diagnostics(SUCCESS_RESULT, edited_file_paths, diagnostics_snapshot)
+        with self.DiagnosticsContext(self, relative_path) as diagnostics_context:
+            self.project.validate_relative_path(relative_path, require_not_ignored=require_not_ignored)
+            with EditedFileContext(relative_path, self.create_code_editor()) as context:
+                original_content = context.get_original_content()
+                replacer = ContentReplacer(mode=mode, allow_multiple_occurrences=allow_multiple_occurrences)
+                updated_content = replacer.replace(original_content, needle, repl)
+                context.set_updated_content(updated_content)
+            return diagnostics_context.format_result(SUCCESS_RESULT)
 
 
 class DeleteLinesTool(EditingToolWithDiagnostics, ToolMarkerOptional):
@@ -255,15 +246,10 @@ class DeleteLinesTool(EditingToolWithDiagnostics, ToolMarkerOptional):
         :param start_line: the 0-based index of the first line to be deleted
         :param end_line: the 0-based index of the last line to be deleted
         """
-        # capturing diagnostics before the edit
-        edited_file_paths = [EditedFilePath(relative_path, relative_path)]
-        diagnostics_snapshot = self._capture_published_lsp_diagnostics_snapshot(edited_file_paths)
-
-        # applying the edit
-        code_editor = self.create_code_editor()
-        code_editor.delete_lines(relative_path, start_line, end_line)
-
-        return self._format_lsp_edit_result_with_new_diagnostics(SUCCESS_RESULT, edited_file_paths, diagnostics_snapshot)
+        with self.DiagnosticsContext(self, relative_path) as diagnostics_context:
+            code_editor = self.create_code_editor()
+            code_editor.delete_lines(relative_path, start_line, end_line)
+            return diagnostics_context.format_result(SUCCESS_RESULT)
 
 
 class ReplaceLinesTool(EditingToolWithDiagnostics, ToolMarkerOptional):
@@ -292,16 +278,12 @@ class ReplaceLinesTool(EditingToolWithDiagnostics, ToolMarkerOptional):
         if not content.endswith("\n"):
             content += "\n"
 
-        # capturing diagnostics before the edit
-        edited_file_paths = [EditedFilePath(relative_path, relative_path)]
-        diagnostics_snapshot = self._capture_published_lsp_diagnostics_snapshot(edited_file_paths)
+        with self.DiagnosticsContext(self, relative_path) as diagnostics_context:
+            code_editor = self.create_code_editor()
+            code_editor.delete_lines(relative_path, start_line, end_line)
+            code_editor.insert_at_line(relative_path, start_line, content)
 
-        # applying the replacement
-        code_editor = self.create_code_editor()
-        code_editor.delete_lines(relative_path, start_line, end_line)
-        code_editor.insert_at_line(relative_path, start_line, content)
-
-        return self._format_lsp_edit_result_with_new_diagnostics(SUCCESS_RESULT, edited_file_paths, diagnostics_snapshot)
+            return diagnostics_context.format_result(SUCCESS_RESULT)
 
 
 class InsertAtLineTool(EditingToolWithDiagnostics, ToolMarkerOptional):
@@ -329,15 +311,11 @@ class InsertAtLineTool(EditingToolWithDiagnostics, ToolMarkerOptional):
         if not content.endswith("\n"):
             content += "\n"
 
-        # capturing diagnostics before the edit
-        edited_file_paths = [EditedFilePath(relative_path, relative_path)]
-        diagnostics_snapshot = self._capture_published_lsp_diagnostics_snapshot(edited_file_paths)
+        with self.DiagnosticsContext(self, relative_path) as diagnostics_context:
+            code_editor = self.create_code_editor()
+            code_editor.insert_at_line(relative_path, line, content)
 
-        # applying the insertion
-        code_editor = self.create_code_editor()
-        code_editor.insert_at_line(relative_path, line, content)
-
-        return self._format_lsp_edit_result_with_new_diagnostics(SUCCESS_RESULT, edited_file_paths, diagnostics_snapshot)
+            return diagnostics_context.format_result(SUCCESS_RESULT)
 
 
 class SearchForPatternTool(Tool):
